feat: add backend/chain transactions synchronisation.

+++ src/main/java/logion/backend/model/transaction/TransactionDescription.java
@@ -1,5 +1,6 @@
 package logion.backend.model.transaction;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import logion.backend.annotation.ValueObject;
 import logion.backend.model.Ss58Address;
@@ -13,10 +14,10 @@ public class TransactionDescription {
 
     Ss58Address from;
     Ss58Address to;
-    long transferValue;
-    long tip;
-    long fee;
-    long reserved;
+    BigInteger transferValue;
+    BigInteger tip;
+    BigInteger fee;
+    BigInteger reserved;
     String pallet;
     String method;
     LocalDateTime createdOn;
